refactored API folder structure
